msm8956: Copy mbn_ota
 * Enable start_ota_daemon prop

+++ rootdir/etc/init.qcom.sh
@@ -110,11 +110,15 @@ if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_versio
     rm -rf /data/misc/radio/modem_config
     mkdir /data/misc/radio/modem_config
     chmod 770 /data/misc/radio/modem_config
+if [ -f /system/etc/mbn_ota.txt ]; then
+    cp /system/etc/mbn_ota.txt /data/misc/radio/modem_config
+else
+    cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
+fi
+    chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
     cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
     chown -hR radio.radio /data/misc/radio/modem_config
     cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
     chown radio.radio /data/misc/radio/ver_info.txt
 fi
-cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
-chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
 echo 1 > /data/misc/radio/copy_complete
